Add files via upload
: This service utilizes an inbuilt map that is connected with other services in 
the building. It gives back the availability of a common space that is shared for team 
meetings, conferences, seminars and other business related activities. If occupied, the 
system will notify interested user on how long is remaining before space is available. If 
unoccupied, the system will notify user of maximum occupancy at the time, and for how 
long the space is available to them.