Add files via upload
This bi-directional streaming service emulates a hypothetical smart office function. It utilizes an inbuilt map that is connected with other services in 
the building. It allows multiple requests and gives back the availability of a common space that is shared for team 
meetings, conferences, seminars and other business related activities.